Move a load of code out of MessageListener:  Separate commands into their own class  Shuffle utils to avoid re-creating work
