--- src/main/java/de/blackcraze/grb/commands/AbstractCommand.java
@@ -1,11 +1,0 @@
-package de.blackcraze.grb.commands;
-
-import java.util.List;
-
-public abstract class AbstractCommand {
-
-	abstract List<String> getChannels();
-
-	abstract boolean trigger();
-
-}
