+++ src/main/java/de/blackcraze/grb/core/Speaker.java
@@ -1,5 +1,6 @@
-package de.blackcraze.grb.commands;
+package de.blackcraze.grb.core;
 
+import net.dv8tion.jda.core.entities.Message;
 import net.dv8tion.jda.core.entities.TextChannel;
 
 public class Speaker {
@@ -14,6 +15,11 @@ public class Speaker {
 			System.err.printf("Can not talk in: %s/%s%n", textChannel.getGuild().getName(), textChannel.getName());
 		}
 		textChannel.sendMessage(text).queue();
+	}
+
+	public static void err(Message message, String text) {
+		message.addReaction(Reaction.FAILURE).queue();
+		say(message.getTextChannel(), text);
 	}
 
 }
