+++ src/main/java/de/blackcraze/grb/dao/StockDaoBean.java
@@ -1,19 +1,18 @@
 package de.blackcraze.grb.dao;
 
-import java.util.List;
-
 import de.blackcraze.grb.model.entity.Mate;
 import de.blackcraze.grb.model.entity.Stock;
 import de.blackcraze.grb.model.entity.StockType;
+
+import java.util.List;
 
 public class StockDaoBean extends BaseDaoBean<Stock> implements IStockDao {
 
 	@SuppressWarnings("unchecked")
 	@Override
 	public List<Stock> findStocksByMate(Mate mate) {
-		StringBuilder b = new StringBuilder();
-		b.append("select s from Stock s where s.mate = :mate order by s.type.name");
-		return em.createQuery(b.toString()).setParameter("mate", mate).getResultList();
+		return em.createQuery("select s from Stock s where s.mate = :mate order by s.type.name")
+				.setParameter("mate", mate).getResultList();
 	}
 
 	@SuppressWarnings("unchecked")
