+++ src/main/java/de/blackcraze/grb/model/Identifiable.java
@@ -1,6 +1,3 @@
-/**
- *
- */
 package de.blackcraze.grb.model;
 
 import java.io.Serializable;
@@ -9,8 +6,5 @@ public interface Identifiable extends Serializable {
 
 	String ID = "id";
 
-	/**
-	 *
-	 */
 	Serializable getId();
 }
