--- src/main/java/de/blackcraze/grb/util/Utils.java
@@ -1,20 +1,0 @@
-package de.blackcraze.grb.util;
-
-import de.blackcraze.grb.core.BotConfig;
-import org.apache.commons.lang3.StringUtils;
-
-import net.dv8tion.jda.core.events.message.guild.GuildMessageReceivedEvent;
-
-public class Utils {
-
-	public static final boolean botMentioned(GuildMessageReceivedEvent event) {
-		return StringUtils.startsWithIgnoreCase(event.getMessage().getContent(), BotConfig.PREFIX);
-		// for (Role role : event.getMessage().getMentionedRoles()) {
-		// if (role.getName().equalsIgnoreCase("bot")) {
-		// return true;
-		// }
-		// }
-		// return false;
-	}
-
-}
